[Improve] Compatibility improvement for parsing Flink config file
